test: add coverage for primer validation hooks
