+++ tsconfig.json
@@ -12,6 +12,8 @@
     "jsx": "react-jsx",
     "noEmit": true,
     "allowJs": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
     "types": ["vitest/globals"]
   },
   "include": ["src"],
